research-companion-final/requirements.txt
requests>=2.31.0
fastapi>=0.95.2
uvicorn>=0.22.0
pydantic>=1.10.11
python-dotenv>=1.0.0
tqdm>=4.66.1
PyMuPDF>=1.22.0
openai>=1.8.0
tiktoken>=0.4.0
sqlalchemy>=2.1.0
pytest>=7.4.2
python-multipart>=0.0.6
